Unify and slightly reword account deletion
The `DeletionCard` was duplicated across the `Profile` for self hosted users, as well as the `Admin` for SaaS.

This now uses the same component for both, and slightly rewords the text.
