--- src/pages/AccountSettings/tabs/Profile/DeletionCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DeletionCard'
